feat(handoff): add @plan/@report/@audit semantic aliases
Add semantic aliases for handoff show command to resolve plan/report/audit
refs from flow_state without needing to know the exact file path.

Changes:
- Add _resolve_artifact_alias() helper in path_helpers.py
- Wire alias resolution into _resolve_shared_artifact()
- Update help text in handoff_read.py with new alias formats
- Add 6 test cases covering all alias scenarios

Users can now use:
- `vibe3 handoff show @plan` to view the plan for current branch
- `vibe3 handoff show @report --branch task/issue-476` to view report
- `vibe3 handoff show @audit` to view the audit for current branch

Refs: #1030
